// src/generated/relay.ts
export interface RelayEnvelope {
    /**
     * @generated from protobuf field: string relay_id = 1
     */
    relayId: string;
    /**
     * @generated from protobuf field: notifier.relay.v1.RelayKind kind = 2
     */
    kind: RelayKind;
    /**
     * @generated from protobuf field: notifier.client.v1.SessionRef sender_session = 3
     */
    senderSession?: SessionRef;
    /**
     * @generated from protobuf field: notifier.client.v1.SessionRef target_session = 4
     */
    targetSession?: SessionRef;
    /**
     * @generated from protobuf field: uint64 seq = 5
     */
    seq: string;
    /**
     * @generated from protobuf field: uint64 ack_seq = 6
     */
    ackSeq: string;
    /**
     * @generated from protobuf field: bytes payload = 7
     */
    payload: Uint8Array;
    /**
     * @generated from protobuf field: int64 sent_at_ms = 8
     */
    sentAtMs: string;
}
/**
 * @generated from protobuf enum notifier.relay.v1.RelayKind
 */
export enum RelayKind {
    /**
     * @generated from protobuf enum value: RELAY_KIND_UNSPECIFIED = 0;
     */
    UNSPECIFIED = 0,
    /**
     * @generated from protobuf enum value: RELAY_KIND_OPEN = 1;
     */
    OPEN = 1,
    /**
     * @generated from protobuf enum value: RELAY_KIND_OPEN_ACK = 2;
     */
    OPEN_ACK = 2,
    /**
     * @generated from protobuf enum value: RELAY_KIND_DATA = 3;
     */
    DATA = 3,
    /**
     * @generated from protobuf enum value: RELAY_KIND_ACK = 4;
     */
    ACK = 4,
    /**
     * @generated from protobuf enum value: RELAY_KIND_CLOSE = 5;
     */
    CLOSE = 5,
    /**
     * @generated from protobuf enum value: RELAY_KIND_PING = 6;
     */
    PING = 6,
    /**
     * @generated from protobuf enum value: RELAY_KIND_ERROR = 7;
     */
    ERROR = 7
}
// ...
export const RelayEnvelope = new RelayEnvelope$Type();